sniptext_chatbot/sniptext_chatbot/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
beautifulsoup4
lxml
requests